requirements.txt
ctranslate2>=4.0,<5
huggingface_hub>=0.13
tokenizers>=0.13,<1
onnxruntime>=1.14,<2 
av>=11
tqdm
python-Levenshtein>=0.21.0
pyaudio
pyannote.audio
python-dotenv
numpy
faster-whisper
torch